move legacy catapp features
