+++ catlearn/fingerprint/adsorbate.py
@@ -115,6 +115,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = make_labels(self.slab_params, '', '_bulk')
         labels.append('ground_state_magmom_bulk')
@@ -144,10 +150,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
 
         Returns
         ----------
-        result : list
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         extra_ads_params = ['atomic_radius', 'heat_of_formation',
                             'oxistates', 'block', 'econf', 'ionenergies']
@@ -175,6 +183,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = make_labels(self.slab_params, '', '_site_av')
         labels.append('ground_state_magmom_site_av')
@@ -196,6 +210,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = make_labels(self.slab_params, '', '_site_min')
         labels.append('ground_state_magmom_site_min')
@@ -217,6 +237,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = make_labels(self.slab_params, '', '_site_max')
         labels.append('ground_state_magmom_site_max')
@@ -238,6 +264,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = make_labels(self.slab_params, '', '_site_med')
         labels.append('ground_state_magmom_site_med')
@@ -258,7 +290,13 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
 
         Parameters
         ----------
-            atoms : object
+        atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = make_labels(self.slab_params, '', '_site_sum')
         labels.append('ground_state_magmom_site_sum')
@@ -279,6 +317,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         if atoms is None:
             return ['av_cn_site', 'cn_site', 'gcn_site', 'cn_ads1', 'gcn_ads1']
@@ -339,18 +383,23 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         """Count the number of neighbors of the site, which has a n number of
         neighbors. This is equivalent to a bag of coordination numbers over the
         site neighbors.
-        These can be used in the "alpha parameters" linear
-        model.
+        These can be used in the "alpha parameters" linear model.
         
         Please cite:
         Roling LT, Abild‐Pedersen F. 
-        Structure‐Sensitive Scaling Relations: 
+        Structure-Sensitive Scaling Relations:
         Adsorption Energies from Surface Site Stability.
         ChemCatChem. 2018 Apr 9;10(7):1643-50.
 
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = ['count_' + str(j) + 'nn_site' for j in range(3, 31)]
         if atoms is None:
@@ -378,6 +427,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         # range of element types.
         raise NotImplementedError("In development.")
@@ -390,6 +445,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         if atoms is None:
             try:
@@ -415,6 +476,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         if atoms is None:
             labels = ['bag_chemi_nn_' + chemical_symbols[z] for z in
@@ -445,6 +512,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = ['nn_surf_ligands', 'identnn_surf_ligands']
         labels += make_labels(self.slab_params, '', '_surf_ligands')
@@ -470,6 +543,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         ads_params = default_params + ['econf', 'ionenergies']
         labels = make_labels(ads_params, '', '_ads_sum')
@@ -491,6 +570,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         ads_params = default_params + ['econf', 'ionenergies']
         labels = make_labels(ads_params, '', '_ads_av')
@@ -512,6 +597,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         if atoms is None:
             return ['strain_site', 'strain_term']
@@ -561,6 +652,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         # range of element types
         n_elements = len(self.atom_types)
@@ -603,6 +700,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         # range of element types
         n_elements = len(self.atom_types)
@@ -646,6 +749,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = ['dist_' + s + '_ads' for s in electronegativities]
         if atoms is None:
@@ -668,6 +777,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = ['dist_' + s + '_chemi' for s in electronegativities]
         if atoms is None:
@@ -693,6 +808,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = ['dist_' + s + '_active' for s in electronegativities]
         if atoms is None:
@@ -712,169 +833,6 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         assert len(en_result) == len(labels)
         return en_result
 
-    def catapp_AB(self, atoms=None):
-        if atoms is None:
-            return ['atomic_number_m1',
-                    'atomic_volume_m1',
-                    'boiling_point_m1',
-                    'density_m1',
-                    'dipole_polarizability_m1',
-                    'electron_affinity_m1',
-                    'group_id_m1',
-                    'lattice_constant_m1',
-                    'melting_point_m1',
-                    'period_m1',
-                    'vdw_radius_m1',
-                    'covalent_radius_cordero_m1',
-                    'en_allen_m1',
-                    'atomic_weight_m1',
-                    'heat_of_formation_m1',
-                    #'dft_bulk_modulus_m1',
-                    #'dft_rhodensity_m1',
-                    #'dbcenter_m1',
-                    #'dbfilling_m1',
-                    #'dbwidth_m1',
-                    #'dbskew_m1',
-                    #'dbkurtosis_m1',
-                    'sblock_m1',
-                    'pblock_m1',
-                    'dblock_m1',
-                    'fblock_m1',
-                    'ne_outer_m1',
-                    'ne_s_m1',
-                    'ne_p_m1',
-                    'ne_d_m1',
-                    'ne_f_m1',
-                    'ionenergy_m1',
-                    'ground_state_magmom_m1',
-                    'atomic_number_m2',
-                    'atomic_volume_m2',
-                    'boiling_point_m2',
-                    'density_m2',
-                    'dipole_polarizability_m2',
-                    'electron_affinity_m2',
-                    'group_id_m2',
-                    'lattice_constant_m2',
-                    'melting_point_m2',
-                    'period_m2',
-                    'vdw_radius_m2',
-                    'covalent_radius_cordero_m2',
-                    'en_allen_m2',
-                    'atomic_weight_m2',
-                    'heat_of_formation_m2',
-                    #'dft_bulk_modulus_m2',
-                    #'dft_rhodensity_m2',
-                    #'dbcenter_m2',
-                    #'dbfilling_m2',
-                    #'dbwidth_m2',
-                    #'dbskew_m2',
-                    #'dbkurtosis_m1',
-                    'sblock_m2',
-                    'pblock_m2',
-                    'dblock_m2',
-                    'fblock_m2',
-                    'ne_outer_m2',
-                    'ne_s_m2',
-                    'ne_p_m2',
-                    'ne_d_m2',
-                    'ne_f_m2',
-                    'ionenergy_m2',
-                    'ground_state_magmom_m2',
-                    'atomic_number_sum',
-                    'atomic_volume_sum',
-                    'boiling_point_sum',
-                    'density_sum',
-                    'dipole_polarizability_sum',
-                    'electron_affinity_sum',
-                    'group_id_sum',
-                    'lattice_constant_sum',
-                    'melting_point_sum',
-                    'period_sum',
-                    'vdw_radius_sum',
-                    'covalent_radius_cordero_sum',
-                    'en_allen_sum',
-                    'atomic_weight_sum',
-                    'heat_of_formation_sum',
-                    #'dft_bulk_modulus_sum',
-                    #'dft_rhodensity_sum',
-                    #'dbcenter_sum',
-                    #'dbfilling_sum',
-                    #'dbwidth_sum',
-                    #'dbskew_sum',
-                    #'dbkurtosis_sum',
-                    'sblock_sum',
-                    'pblock_sum',
-                    'dblock_sum',
-                    'fblock_sum',
-                    'ne_outer_sum',
-                    'ne_s_sum',
-                    'ne_p_sum',
-                    'ne_d_sum',
-                    'ne_f_sum',
-                    'ionenergy_sum',
-                    'ground_state_magmom_sum',
-                    'concentration_catapp',
-                    'facet_catapp',
-                    'site_catapp']
-        else:
-            # Atomic numbers in the site.
-            Z_surf1_raw = [atoms.numbers[j]
-                           for j in atoms.subsets['ligand_atoms']]
-            # Sort by concentration
-            counts = collections.Counter(Z_surf1_raw)
-            Z_surf1 = sorted(Z_surf1_raw, key=counts.get, reverse=True)
-            z1 = Z_surf1[0]
-            z2 = Z_surf1[0]
-            for z in Z_surf1:
-                if z != z1:
-                    z2 = z
-            uu, ui = np.unique(Z_surf1, return_index=True)
-            if len(ui) == 1:
-                if Z_surf1[0] == z1:
-                    site = 1.
-                elif Z_surf1[0] == z2:
-                    site = 3.
-            else:
-                site = 2.
-            # Import overlayer composition from ase database.
-            kvp = atoms.info['key_value_pairs']
-            term = [atomic_numbers[zt] for zt in string2symbols(kvp['term'])]
-            termuu, termui = np.unique(term, return_index=True)
-            if '3' in kvp['term']:
-                conc = 3.
-            elif len(termui) == 1:
-                conc = 1.
-            elif len(termui) == 2:
-                conc = 2.
-            else:
-                raise NotImplementedError("catappAB only supports AxBy.")
-            text_params = default_params + ['heat_of_formation',
-                                            #'dft_bulk_modulus',
-                                            #'dft_density',
-                                            #'dbcenter',
-                                            #'dbfilling',
-                                            #'dbwidth',
-                                            #'dbskew',
-                                            #'dbkurt',
-                                            'block',
-                                            'econf',
-                                            'ionenergies']
-            f1 = get_mendeleev_params(z1, params=text_params)
-            f1 = f1[:-3] + block2number[f1[-3]] + \
-                list(n_outer(f1[-2])) + [f1[-1]['1']] + \
-                [float(gs_magmom[z1])]
-            if z1 == z2:
-                f2 = f1
-            else:
-                f2 = get_mendeleev_params(z2, params=text_params)
-                f2 = f2[:-3] + block2number[f2[-3]] + \
-                    list(n_outer(f2[-2])) + [f2[-1]['1']] + \
-                    [float(gs_magmom[z2])]
-            msum = list(np.nansum([f1, f2], axis=0, dtype=np.float))
-            facet = facetdict[kvp['facet'].replace(')', '').replace('(', '')]
-            fp = f1 + f2 + msum + [conc] + facet + [site]
-            return fp
-
     def delta_energy(self, atoms=None):
         """Return the contents of
         atoms.info['key_value_pairs']['delta_energy'] as a feature.
@@ -882,6 +840,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         if atoms is None:
             return ['delta_energy']
@@ -899,7 +863,13 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
 
         Parameters
         ----------
-            atoms : object
+        atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         labels = ['layers', 'size', 'coverage']
         if atoms is None:
@@ -917,27 +887,18 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
             coverage = np.true_divide(n, size)
             return layers, size, coverage
 
-    def name(self, atoms=None):
-        """Return a name for a datapoint based on the contents of
-        atoms.info['key_value_pairs'].
-
-        Parameters
-        ----------
-        atoms : object
-        """
-        if atoms is None:
-            return ['catapp_name']
-        else:
-            kvp = atoms.info['key_value_pairs']
-            name = kvp['species'] + '*' + kvp['name'] + kvp['facet']
-            return [name]
-
     def dbid(self, atoms=None):
         """Return the contents of atoms.info['id'] as a feature.
 
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         if atoms is None:
             return ['id']
@@ -952,6 +913,12 @@ class AdsorbateFingerprintGenerator(BaseGenerator):
         Parameters
         ----------
         atoms : object
+            ASE Atoms object.
+
+        Returns
+        ----------
+        features : list
+            If None was passed, the elements are strings, naming the feature.
         """
         if atoms is None:
             return ['time_float']
